Users can no longer add unsupported links into a playlist

+++ discord-music-bot/src/main/java/net/masterzach32/swagbot/App.java
@@ -439,8 +439,12 @@ public class App {
                 playlist.toggleRequiresPerms();
                 response = playlist.requiresPerms() ? "Playlist **" + playlist.getName() + "** now requires moderator privelages to edit." : "Playlist **" + playlist.getName() + "** no longer requires moderator privelages to edit.";
             } else if (command.equals("add") && !(playlist.requiresPerms() && !perms) && !playlist.isLocked()) {
-                message.delete();
-                response = playlist.add(params[2]) ? "Added **" + new YouTubeAudio(params[2]).getTitle() + "** to **" + playlist.getName() + "**" : "Playlist **" + playlist.getName() + "** already has " + new YouTubeAudio(params[2]).getTitle();
+                if(!params[2].contains("youtube") && !params[2].contains("soundcloud"))
+                    response = "You must provide a YouTube or SoundCloud link!";
+                else {
+                    message.delete();
+                    response = playlist.add(params[2]) ? "Added **" + new YouTubeAudio(params[2]).getTitle() + "** to **" + playlist.getName() + "**" : "Playlist **" + playlist.getName() + "** already has " + new YouTubeAudio(params[2]).getTitle();
+                }
             } else if (command.equals("remove") && perms && !playlist.isLocked()) {
                 playlist.remove(params[2]);
                 response = "Removed " + params[2] + " from **" + playlist.getName() + "**";
